remove Expo template boilerplate and simplify root layout
Replace tab navigation, themed components, hooks, and example screens
with a minimal single-screen setup using a plain Stack navigator.
