--- components/hello-wave.tsx
@@ -1,20 +1,0 @@
-import Animated from "react-native-reanimated";
-
-export function HelloWave() {
-  return (
-    <Animated.Text
-      style={{
-        fontSize: 28,
-        lineHeight: 32,
-        marginTop: -6,
-        animationName: {
-          "50%": { transform: [{ rotate: "25deg" }] },
-        },
-        animationIterationCount: 4,
-        animationDuration: "300ms",
-      }}
-    >
-      👋
-    </Animated.Text>
-  );
-}
